requirements.txt
pytubefix>=0.6.0
colorama>=0.4.6
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=6.0.0
